2025-05-27/requirements.txt
streamlit>=1.24.0
langchain>=0.0.300
langgraph>=0.0.10
openai>=1.0.0
langsmith>=0.0.50
chromadb>=0.4.0
extract-msg>=0.41.0
python-daemon>=3.0.1
schedule>=1.2.0
lockfile>=0.12.2
python-dotenv>=1.0.0
pytest==8.3.5
schedule==1.2.1
PyPDF2>=3.0.0
psutil>=5.9.0
sentence-transformers>=2.2.0 